# docs/swagger.yaml
definitions:
...
  models.ChangePasswordInput:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  models.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  models.HttpResponseError:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.HttpResponseOK:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.ResetPasswordInput:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  models.SendVerificationInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  models.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - password
    type: object
  models.UserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        items:
          type: string
        type: array
      surname:
        type: string
      user_role:
        type: string
    type: object
info:
  contact:
    email: proteng.plus@gmail.com
    name: Proteng API Support
  title: Proteng API
  version: "1.0"
paths:
...
  /proteng-user-mgmt/auth/changepassword:
    patch:
      consumes:
      - application/json
      description: Changes the password while logged in
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: change password
      tags:
      - Authentication
  /proteng-user-mgmt/auth/forgotpassword:
    post:
      consumes:
      - application/json
      description: Initiates the forgot password process by sending a reset email
        to the user
      parameters:
      - description: User email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Initiate forgot password process
      tags:
      - Authentication
  /proteng-user-mgmt/auth/login:
    post:
      consumes:
      - application/json
      description: Signs in a user by validating the provided credentials
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Sign in a user
      tags:
      - Authentication
...
  /proteng-user-mgmt/auth/resetpassword/{resetToken}:
    patch:
      consumes:
      - application/json
      description: Resets the password using the provided reset token
      parameters:
      - description: Reset Token
        in: path
        name: resetToken
        required: true
        type: string
      - description: New password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Reset password
      tags:
      - Authentication
  /proteng-user-mgmt/auth/sentverification:
    post:
      consumes:
      - application/json
      description: Initiates the email verification process by sending a verification
        email to the user
      parameters:
      - description: User email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SendVerificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: initiates verification process
      tags:
      - Authentication
  /proteng-user-mgmt/auth/verifyemail/{verificationToken}:
    patch:
      consumes:
      - application/json
      description: Verifies the email using the provided verification token
      parameters:
      - description: Verification Token
        in: path
        name: verificationToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Verify email
      tags:
      - Authentication
  /proteng-user-mgmt/me:
    delete:
      description: Delete the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete the currently authenticated user
      tags:
      - User management
    get:
      description: Retrieve details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - User management
    put:
      consumes:
      - application/json
      description: Update details of the currently authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update the currently authenticated user
      tags:
      - User management
  /proteng-user-mgmt/users:
    get:
      description: Retrieve a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve all users
      tags:
      - User management
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      summary: Create a new user
      tags:
      - User management
  /proteng-user-mgmt/users/{id}:
    delete:
      description: Delete a user by providing its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - User management
    get:
      description: Retrieve a user by providing its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a user by ID
      tags:
      - User management
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.HttpResponseOK'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.HttpResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update an existing user
      tags:
      - User management
securityDefinitions:
  ApiKeyAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"